``` feat: Add admin HTMX fragments and enhance businesses API with UUID support and HTMX handlers
## Controller Changes (C)
- handlers/admin.rs: Add HTMX fragments for moderation, analytics, and audit logs
  - admin_moderation_fragment: Render moderation queue with status filter, priority badges, approve/reject actions
  - admin_analytics_fragment: Paginated analytics events table with user/community context
  - admin_audit_logs_fragment: Paginated audit log timeline with action details
  - Add

+++ src/server/src/handlers/proposals.rs
@@ -149,7 +149,7 @@ pub async fn list_proposals(
                 .map(|dt| dt.to_rfc3339()),
             voting_ends_at: row.get::<Option<chrono::DateTime<chrono::Utc>>, _>("voting_ends_at")
                 .map(|dt| dt.to_rfc3339()),
-            quorum_required: row.get::<Option<i64>, _>("quorum_required").unwrap_or(0) as i32,
+            quorum_required: row.get::<Option<i32>, _>("quorum_required").unwrap_or(0),
         }
     }).collect();
     
@@ -196,7 +196,7 @@ pub async fn get_proposal(
             .map(|dt| dt.to_rfc3339()),
         voting_ends_at: row.get::<Option<chrono::DateTime<chrono::Utc>>, _>("voting_ends_at")
             .map(|dt| dt.to_rfc3339()),
-        quorum_required: row.get::<Option<i64>, _>("quorum_required").unwrap_or(0) as i32,
+        quorum_required: row.get::<Option<i32>, _>("quorum_required").unwrap_or(0),
     }))
 }
 
@@ -377,7 +377,7 @@ pub async fn get_results(
     .map_err(|e| AppError::Internal(anyhow::anyhow!("Database error: {}", e)))?
     .ok_or_else(|| AppError::Internal(anyhow::anyhow!("Proposal not found".to_string())))?;
     
-    let quorum_required: i64 = proposal.get::<Option<i64>, _>("quorum_required").unwrap_or(0);
+    let quorum_required: i64 = proposal.get::<Option<i32>, _>("quorum_required").unwrap_or(0) as i64;
     let community_id: Uuid = proposal.get("community_id");
     
     // Get vote counts by value
